refactor: ios 푸시알림 코드 리펙토링
