--- backend/src/main/java/com/dongguk/cse/naemansan/controller/FCMNotificationApiController.java
@@ -1,21 +1,0 @@
-package com.dongguk.cse.naemansan.controller;
-
-import com.dongguk.cse.naemansan.dto.request.FCMNotificationRequestDto;
-import com.dongguk.cse.naemansan.service.FCMNotificationService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/v1/notification")
-public class FCMNotificationApiController {
-    private final FCMNotificationService fcmNotificationService;
-
-    @PostMapping()
-    public String sendNotificationByToken(@RequestBody FCMNotificationRequestDto requestDto) {
-        return fcmNotificationService.sendNotificationByToken(requestDto);
-    }
-}
