--- backend/src/main/java/com/dongguk/cse/naemansan/service/FirebaseCloudMessageService.java
@@ -1,73 +1,0 @@
-package com.dongguk.cse.naemansan.service;
-
-import com.dongguk.cse.naemansan.dto.MessageDto;
-//import com.dongguk.cse.naemansan.dto.NotificationDto;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.gson.JsonParseException;
-
-import lombok.RequiredArgsConstructor;
-import okhttp3.*;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.stereotype.Service;
-
-import java.io.IOException;
-import java.util.List;
-
-
-@Service
-@RequiredArgsConstructor
-public class FirebaseCloudMessageService {
-    private final String API_URL = "https://fcm.googleapis.com/v1/projects/wearebility-303e9/messages:send";
-    private final ObjectMapper objectMapper;
-
-    public void sendMessageTo(String targetToken, String title, String body) throws IOException {
-        //System.out.println("sendMessageTo 시작");
-        String message = makeMessage(targetToken, title, body);
-
-        OkHttpClient client = new OkHttpClient();
-        RequestBody requestBody = RequestBody.create(message, MediaType.get("application/json; charset=utf-8"));
-        Request request = new Request.Builder()
-                .url(API_URL)
-                .post(requestBody)
-                .addHeader(HttpHeaders.AUTHORIZATION, "Bearer " + getAccessToken())
-                .addHeader(HttpHeaders.CONTENT_TYPE, "application/json; UTF-8")
-                .build();
-
-        Response response = client.newCall(request).execute();
-        //System.out.println("sendMessageTo 끝나기 전");
-        System.out.println(response.body().string());
-    }
-
-    private String makeMessage(String targetToken, String title, String body) throws JsonParseException, JsonProcessingException {
-        //System.out.println("makeMessage 시작");
-
-        MessageDto messageDto = MessageDto.builder()
-                .message(MessageDto.Message.builder()
-                        .token(targetToken)
-                        .notification(MessageDto.Notification.builder()
-                                .title(title)
-                                .body(body)
-                                .image(null)
-                                .build()
-                        ).build()).validateOnly(false).build();
-
-
-        System.out.println(targetToken + " / " + title + " / " + body);
-        //System.out.println("makeMessage 끝");
-        return objectMapper.writeValueAsString(messageDto);
-    }
-
-    private String getAccessToken() throws IOException {
-        String firebaseConfigPath = "firebase/firebase_service_key.json";
-
-        GoogleCredentials googleCredentials = GoogleCredentials
-                .fromStream(new ClassPathResource(firebaseConfigPath).getInputStream())
-                .createScoped(List.of("https://www.googleapis.com/auth/cloud-platform"));
-
-        googleCredentials.refreshIfExpired();
-        return googleCredentials.getAccessToken().getTokenValue();
-    }
-}
