+++ backend/src/main/java/com/dongguk/cse/naemansan/service/NotificationService.java
@@ -10,9 +10,20 @@ import com.dongguk.cse.naemansan.dto.request.FCMNotificationRequestDto;
 import com.dongguk.cse.naemansan.dto.request.NotificationRequestDto;
 import com.dongguk.cse.naemansan.repository.NotificationRepository;
 import com.dongguk.cse.naemansan.repository.UserRepository;
+import com.dongguk.cse.naemansan.util.NotificationUtil;
+import com.google.firebase.messaging.FirebaseMessaging;
+import com.google.firebase.messaging.FirebaseMessagingException;
+import com.google.firebase.messaging.Message;
 import jakarta.transaction.Transactional;
+import javapns.Push;
+import javapns.communication.exceptions.CommunicationException;
+import javapns.communication.exceptions.KeystoreException;
+import javapns.notification.PushNotificationPayload;
+import javapns.notification.PushedNotification;
+import javapns.notification.ResponsePacket;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.json.JSONException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -33,8 +44,9 @@ public class NotificationService {
     //title 처리 해야함
     private final UserRepository userRepository;
     private final NotificationRepository notificationRepository;
-    private final FirebaseCloudMessageService firebaseCloudMessageService;
-    private final FCMNotificationService fcmNotificationService;
+   // private final FCMNotificationService fcmNotificationService;
+    private final FirebaseMessaging firebaseMessaging;
+    NotificationUtil notificationUtil;
 
     //NotificationDto 삭제
     public /*ResponseEntity*/String createNotification(Long userId, FCMNotificationRequestDto requestDto /*NotificationRequestDto notificationRequestDto*/) throws IOException {
@@ -46,7 +58,7 @@ public class NotificationService {
                 .content(requestDto.getBody())
                 .build());
 
-        return fcmNotificationService.sendNotificationByToken(requestDto);
+        return notificationUtil.sendNotificationByToken(requestDto);
 /*
         firebaseCloudMessageService.sendMessageTo(
                 notificationRequestDto.getTargetToken(),
@@ -98,4 +110,6 @@ public class NotificationService {
         notificationRepository.delete(notification);
         return Boolean.TRUE;
     }
+    //ios 푸시알림 보내기
+
 }
